I am now able to choose format and convert files

+++ Homework Assignments/Final/convert.php
@@ -1,16 +1,4 @@
 #!/usr/local/bin/php
-<?php 
-    //ig get request is made then get file name and download it as photo.png
-    if ($_SERVER['REQUEST_METHOD'] === 'GET'){
-        ob_start();
-        $outputName = $_GET['file']; 
-        $file = "./uploads/" . $outputName;
-        header("Content-Type: application/octet-stream");
-        header('Content-Disposition: attachment; filename="' . $outputName . '"');
-        readfile ($file);
-        exit();
-    }
-?>
 <!DOCTYPE html>
 <html lang="en">
 <head>
@@ -22,30 +10,35 @@
 </head>
 <body>
 <?php 
-    if(isset($_POST['submit'])){
+
+    if(isset($_POST['submit']) && isset($_POST['radio_button'])){
         $fileNames = $_FILES['their_files']['name'];
         $fileTempNames= $_FILES['their_files']['tmp_name'];
+        $fileType = $_POST['radio_button'];
         
         //eader("Content-Type: application/octet-stream");
         for ($x = 0; $x < count($fileNames); $x++) {
             
                 $saveLocation = dirname(realpath(__FILE__)) . '/uploads/' . $fileNames[$x];
                 move_uploaded_file($fileTempNames[$x], $saveLocation);
-                exec("mogrify -format png $saveLocation");
+                exec("mogrify -format $fileType $saveLocation");
                 exec("rm -f $saveLocation");
-                //header('Content-Disposition: attachment; filename="' . $fileNames[$x] . '"');
-                echo "The name is: $fileNames[$x] <br>";
+                //echo "The name is: $fileNames[$x] <br>";
                 $fileExplode = explode('.', $fileNames[$x]);
-                $fileBase = $fileExplode[0].'.png';
+                $fileBase = $fileExplode[0].'.'.$fileType;
                 echo "<script>window.addEventListener('load', ()=>{ transform('$fileBase'); }) ;</script>";
         
         }
+        echo "Congrats! You have successfully downloaded your files!";
 
         // $fileType = $_POST['radio_button'];
         // echo $fileType;
-        echo $_POST['radio_button'];
+        //echo $_POST['radio_button'];
 
     } 
+    else{
+        echo "Choose a file type!";
+    }
 ?>
 </body>
 </html>
